uso de Middlewares

+++ routes/web.php
@@ -11,8 +11,8 @@
 |
 */
 
-Route::get('/', 'UserController@index');
-Route::post('users', 'UserController@store')->name('users.store');
-Route::delete('users/{user}', 'UserController@destroy')->name('users.destroy');
+Route::get('/', 'UserController@index')->middleware('auth');
+Route::post('users', 'UserController@store')->name('users.store')->middleware('auth');
+Route::delete('users/{user}', 'UserController@destroy')->name('users.destroy')->middleware('auth');
